requirements.txt
PyQt5
python-docx
python-pptx
